move ToolOptions into the Config namespace
