+++ spec/quiet_quality/config/tool_options_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe QuietQuality::ToolOptions do
+RSpec.describe QuietQuality::Config::ToolOptions do
   let(:tool) { "foo" }
   subject(:tool_options) { described_class.new(tool) }
